Add ApiMethodParamTransformer
